requirements.txt
apex==0.9.10dev
bert_score==0.3.11
ftfy==6.1.1
matplotlib==3.5.1
numpy==1.21.2
opencv_python==4.5.5.62
Pillow==9.1.0
refer==0.0.1
regex==2022.1.18
ruamel.base==1.0.0
ruamel_yaml==0.17.21
scikit_image==0.18.3
scipy==1.7.3
seaborn==0.11.2
setuptools==58.0.4
skimage==0.0
tensorflow==2.8.0
timm==0.4.9
torch==1.10.2
torchvision==0.11.3
tqdm==4.62.3
transformers==4.8.1
xlwt==1.3.0
